refactor in process, smth happen with DB
