+++ src/main/java/com/jm/online_store/service/impl/ProductServiceImpl.java
@@ -21,12 +21,15 @@ import com.opencsv.bean.CsvToBean;
 import com.opencsv.bean.CsvToBeanBuilder;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.poi.xssf.usermodel.XSSFCell;
 import org.apache.poi.xssf.usermodel.XSSFRow;
 import org.apache.poi.xssf.usermodel.XSSFSheet;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.multipart.MultipartFile;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -37,8 +40,10 @@ import javax.mail.MessagingException;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
+import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.Reader;
 import java.nio.file.Files;
@@ -294,62 +299,62 @@ public class ProductServiceImpl implements ProductService {
      */
     @Override
     public void importFromXMLFile(String fileName) {
-        try {
-            // Создается построитель документа
-            DocumentBuilder documentBuilder = DocumentBuilderFactory.newInstance().newDocumentBuilder();
-
-            // Создается дерево DOM документа из файла
-            Document document = documentBuilder.parse("uploads/import/" + fileName);
-            document.getDocumentElement().normalize();
-
-            NodeList nList = document.getElementsByTagName("product");
-
-            for (int temp = 0; temp < nList.getLength(); temp++) {
-
-                Node nNode = nList.item(temp);
-                if (nNode.getNodeType() == Node.ELEMENT_NODE) {
-
-                    Element eElement = (Element) nNode;
-
-                    String productName = eElement.getElementsByTagName("productname").item(0).getTextContent();
-                    String productPrice = eElement.getElementsByTagName("price").item(0).getTextContent();
-                    String productAmount = eElement.getElementsByTagName("amount").item(0).getTextContent();
-                    String categoryName = eElement.getElementsByTagName("categoryname").item(0).getTextContent();
-                    Product product = new Product(productName, Double.parseDouble(productPrice), Integer.parseInt(productAmount));
-                    categoriesService.addToProduct(product, categoryName);
-
-                    String characteristicName;
-                    String characteristicValue;
-                    if (eElement.getElementsByTagName("characteristicname").getLength() != 0
-                            && eElement.getElementsByTagName("characteristicvalue").getLength() != 0) {
-                        characteristicName = eElement.getElementsByTagName("characteristicname").item(0).getTextContent();
-                        characteristicValue = eElement.getElementsByTagName("characteristicvalue").item(0).getTextContent();
-
-                        List<String> listNames = Arrays.asList(characteristicName.split(","));
-                        List<String> listValues = Arrays.asList(characteristicValue.split(","));
-                        Map<String, String> map = new HashMap<>();
-
-                        for (int i = 0; i < listValues.size(); i++) {
-                            map.put(listNames.get(i), listValues.get(i));
-                        }
-                        for (Map.Entry<String, String> entry : map.entrySet()) {
-                            productCharacteristicService.addProductCharacteristic(findProductByName(productName).orElseThrow(ProductNotFoundException::new).getId(),
-                                    entry.getKey(), entry.getValue());
-                        }
-                    }
-                }
-            }
-
-        } catch (ParserConfigurationException e) {
-            log.error("Ошибка конфигурации парсера");
-            e.printStackTrace();
-        } catch (SAXException e) {
-            log.error("Ошибка XML синтаксиса");
-            e.printStackTrace();
-        } catch (IOException e) {
-            log.error("Ошибка ввода/вывода");
-            e.printStackTrace();
-        }
+//        try {
+//            // Создается построитель документа
+//            DocumentBuilder documentBuilder = DocumentBuilderFactory.newInstance().newDocumentBuilder();
+//
+//            // Создается дерево DOM документа из файла
+//            Document document = documentBuilder.parse("uploads/import/" + fileName);
+//            document.getDocumentElement().normalize();
+//
+//            NodeList nList = document.getElementsByTagName("product");
+//
+//            for (int temp = 0; temp < nList.getLength(); temp++) {
+//
+//                Node nNode = nList.item(temp);
+//                if (nNode.getNodeType() == Node.ELEMENT_NODE) {
+//
+//                    Element eElement = (Element) nNode;
+//
+//                    String productName = eElement.getElementsByTagName("productname").item(0).getTextContent();
+//                    String productPrice = eElement.getElementsByTagName("price").item(0).getTextContent();
+//                    String productAmount = eElement.getElementsByTagName("amount").item(0).getTextContent();
+//                    String categoryName = eElement.getElementsByTagName("categoryname").item(0).getTextContent();
+//                    Product product = new Product(productName, Double.parseDouble(productPrice), Integer.parseInt(productAmount));
+//                    categoriesService.addToProduct(product, categoryName);
+//
+//                    String characteristicName;
+//                    String characteristicValue;
+//                    if (eElement.getElementsByTagName("characteristicname").getLength() != 0
+//                            && eElement.getElementsByTagName("characteristicvalue").getLength() != 0) {
+//                        characteristicName = eElement.getElementsByTagName("characteristicname").item(0).getTextContent();
+//                        characteristicValue = eElement.getElementsByTagName("characteristicvalue").item(0).getTextContent();
+//
+//                        List<String> listNames = Arrays.asList(characteristicName.split(","));
+//                        List<String> listValues = Arrays.asList(characteristicValue.split(","));
+//                        Map<String, String> map = new HashMap<>();
+//
+//                        for (int i = 0; i < listValues.size(); i++) {
+//                            map.put(listNames.get(i), listValues.get(i));
+//                        }
+//                        for (Map.Entry<String, String> entry : map.entrySet()) {
+//                            productCharacteristicService.addProductCharacteristic(findProductByName(productName).orElseThrow(ProductNotFoundException::new).getId(),
+//                                    entry.getKey(), entry.getValue());
+//                        }
+//                    }
+//                }
+//            }
+//
+//        } catch (ParserConfigurationException e) {
+//            log.error("Ошибка конфигурации парсера");
+//            e.printStackTrace();
+//        } catch (SAXException e) {
+//            log.error("Ошибка XML синтаксиса");
+//            e.printStackTrace();
+//        } catch (IOException e) {
+//            log.error("Ошибка ввода/вывода");
+//            e.printStackTrace();
+//        }
     }
 
     /**
@@ -696,5 +701,88 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void deleteProductFromTrackedForLoggedInUser(long productId) {
         productRepository.deletePriceChangeSubscriber(userService.getCurrentLoggedInUser().getEmail(), productId);
+    }
+
+    @Override
+    public void importFromXMLFile(MultipartFile file) {
+        writeFile(file);
+        try {
+            // Создается построитель документа
+            DocumentBuilder documentBuilder = DocumentBuilderFactory.newInstance().newDocumentBuilder();
+            // Создается дерево DOM документа из файла
+            Document document = documentBuilder.parse("C:\\Users\\Dalan\\Desktop\\dev\\online_store\\uploads\\import\\" + file.getOriginalFilename());
+            document.getDocumentElement().normalize();
+            NodeList nList = document.getElementsByTagName("product");
+
+            for (int temp = 0; temp < nList.getLength(); temp++) {
+
+                Node nNode = nList.item(temp);
+                if (nNode.getNodeType() == Node.ELEMENT_NODE) {
+
+                    Element eElement = (Element) nNode;
+
+                    String productName = eElement.getElementsByTagName("productname").item(0).getTextContent();
+                    String productPrice = eElement.getElementsByTagName("price").item(0).getTextContent();
+                    String productAmount = eElement.getElementsByTagName("amount").item(0).getTextContent();
+                    String categoryName = eElement.getElementsByTagName("categoryname").item(0).getTextContent();
+                    Product product = new Product(productName, Double.parseDouble(productPrice), Integer.parseInt(productAmount));
+                    categoriesService.addToProduct(product, categoryName);
+
+                    String characteristicName;
+                    String characteristicValue;
+                    if (eElement.getElementsByTagName("characteristicname").getLength() != 0
+                            && eElement.getElementsByTagName("characteristicvalue").getLength() != 0) {
+                        characteristicName = eElement.getElementsByTagName("characteristicname").item(0).getTextContent();
+                        characteristicValue = eElement.getElementsByTagName("characteristicvalue").item(0).getTextContent();
+
+                        List<String> listNames = Arrays.asList(characteristicName.split(","));
+                        List<String> listValues = Arrays.asList(characteristicValue.split(","));
+                        Map<String, String> map = new HashMap<>();
+
+                        for (int i = 0; i < listValues.size(); i++) {
+                            map.put(listNames.get(i), listValues.get(i));
+                        }
+                        for (Map.Entry<String, String> entry : map.entrySet()) {
+                            productCharacteristicService.addProductCharacteristic(findProductByName(productName).orElseThrow(ProductNotFoundException::new).getId(),
+                                    entry.getKey(), entry.getValue());
+                        }
+                    }
+                }
+            }
+
+        } catch (ParserConfigurationException e) {
+            log.error("Ошибка конфигурации парсера");
+            e.printStackTrace();
+        } catch (SAXException e) {
+            log.error("Ошибка XML синтаксиса");
+            e.printStackTrace();
+        } catch (IOException e) {
+            log.error("Ошибка ввода/вывода");
+            e.printStackTrace();
+        }
+
+    }
+
+    @Override
+    public void importFromCSVFile(MultipartFile multipartFile) throws FileNotFoundException {
+
+    }
+
+
+
+    private void writeFile(@RequestParam("file") MultipartFile file) {
+        try {
+            byte[] bytes = file.getBytes();
+            if (bytes.length == 0)
+                throw new Exception("File is empty");
+            BufferedOutputStream stream =
+                    new BufferedOutputStream(new FileOutputStream("C:\\Users\\Dalan\\Desktop\\dev\\online_store\\uploads\\import\\" + file.getOriginalFilename()));
+            stream.write(bytes);
+            stream.close();
+        } catch (Exception e) {
+            log.error("Ошибка сохранения файла");
+            e.printStackTrace();
+        }
+        log.debug("тип файла" + FilenameUtils.getExtension(file.getOriginalFilename()));
     }
 }
